requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dash==4.1.0
Flask==3.1.3
greenlet==3.3.2
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
plotly==6.6.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.0
retrying==1.4.2
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
ucimlrepo==0.0.7
urllib3==2.6.3
Werkzeug==3.1.7
zipp==3.23.0
